solve comments

+++ src/azureml-studio-model/azureml/studio/model/builtin_models/pytorch/cloudpickle.py
@@ -1,19 +1,21 @@
+import os
+
 import cloudpickle
-import pandas as pd
-import torch
-import torchvision
 
 from .base import PytorchBaseModel
+
 
 class PytorchCloudPickleModel(PytorchBaseModel):
 
     flavor = {
         "name": "pytorch",
-        "serializtion_method": "cloudpickle",
+        "serialization_method": "cloudpickle",
         "is_cuda": False
     }
 
     def save(self, save_to, overwrite_if_exists=True):
+        if os.path.isfile(save_to) and not overwrite_if_exists:
+            raise Exception(f"File {save_to} exists. Set overwrite_is_exists=True if you want to overwrite it.")
         with open(save_to, "wb") as fp:
             cloudpickle.dump(self.raw_model, fp)
